add same to scheduler

+++ main-report.c
@@ -23,7 +23,7 @@ static const char *current_dev = "/sys/class/power_supply/bms/current_now";
 
 static int64_t voltage_sum = 0;
 static int64_t current_sum = 0;
-static int core_num = 0;
+static int64_t core_num = 0;
 static int64_t num_samples = 0;
 
 static struct timespec start_time;
